% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_predict.R
\name{catch_predict}
\alias{catch_predict}
\title{function to predict catch by stock, gear and week based on predictions from an encounter rate model and a bunch of user inputs}
\usage{
catch_predict(
  schedule,
  mort_rates,
  licenses,
  sets_per_day,
  mark_rates,
  run_size,
  weekly_HR,
  adult_props,
  allocation
)
}
\arguments{
\item{allocation}{}
}
\description{
function to predict catch by stock, gear and week based on predictions from an encounter rate model and a bunch of user inputs
}
