% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{download_template}
\alias{download_template}
\title{Download the Excel template}
\usage{
download_template(path = ".", filename = "Inputs.xlsx")
}
\arguments{
\item{path}{Character. The directory where the template should be saved. Defaults to the current working directory.}

\item{filename}{Character. The name of the saved file. Defaults to "template.xlsx".}
}
\value{
The full path of the saved file.
}
\description{
Copies the Excel template file to a specified directory.
}
