% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECF_mod.R
\name{ECF_mod}
\alias{ECF_mod}
\title{ECF_mod}
\usage{
ECF_mod(
  path = system.file("extdata", "Inputs.xlsx", package = "ECFchinook"),
  year = 2025,
  find_quota = TRUE
)
}
\arguments{
\item{path}{Character. The file path to the input Excel file.}

\item{year}{Number. The year to model}
}
\description{
ECF_mod
}
