% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECF_mod.R
\name{ECF_mod_report}
\alias{ECF_mod_report}
\title{Make ECF Chinook quota planning report}
\usage{
ECF_mod_report(
  Input_path = system.file("extdata", "Inputs.xlsx", package = "ECFchinook"),
  year = 2025,
  find_quota = TRUE,
  output_file = "ECF Chinook quota planning report",
  output_dir = getwd(),
  ...
)
}
\arguments{
\item{year}{Number. The year to model}

\item{find_quota}{Boolean. TRUE to find the maximum number of kept Chinook for fishing with the distribution of effort specified in the schedule tab. FALSE to find the expected catch with days of week specified in the schedule tab.}

\item{output_file}{Charachter. The file name for the report.}

\item{output_dir}{The output directory for the report. Defaults to the current working directory.}

\item{...}{}
}
\description{
A call to this function finds the Chinook quotas associated with the values in an input Excel file. It saves an HTML report with information on the inputs and the resulting quotas and expected harvest and impacts. It also saves two CSV files, one with Chinook impact values for use in the pre-season Chinook MR and one for the pre-season steelhead model.
}
