% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs_and_figs_fun.R
\name{weekly_impacts_fig_fun}
\alias{weekly_impacts_fig_fun}
\title{weekly_impacts_fig_fun}
\usage{
weekly_impacts_fig_fun(dat, weekdates)
}
\arguments{
\item{weekdates}{}
}
\description{
weekly_impacts_fig_fun
}
